[gha] Automate release process
Adds the 'release-bitergia-analytics' GitHub workflow
to automate the release of Bitergia Analytics and its
components. It must be triggered manually. The workflow
uses release-tools commands to select the highest possible
semantic version among all components, create release notes
and a tag in all the repositories.
The 'release' workflow is then triggered by the new tag.

Signed-off-by: Eva Millán <evamillan@bitergia.com>